Moved categories to categories module
