+++ frontend/src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from "@angular/router";
 
-import { CategoryListComponent } from "./category-list/category-list.component";
-import { CategoryDetailComponent } from "./category-detail/category-detail.component";
+import { CategoryListComponent } from "./categories/category-list/category-list.component";
+import { CategoryDetailComponent } from "./categories/category-detail/category-detail.component";
 import { FoodComponent } from './food/food.component';
 import { FoodDetailComponent } from './food-detail/food-detail.component';
