+++ frontend/src/app/app.module.ts
@@ -3,11 +3,11 @@ import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import { HttpClientModule } from '@angular/common/http';
-import { CategoryListComponent } from './category-list/category-list.component';
+import { CategoryListComponent } from './categories/category-list/category-list.component';
 import { AppRoutingModule } from './app-routing.module';
 import { FormsModule } from "@angular/forms";
-import { CategoryDetailComponent } from './category-detail/category-detail.component';
-import { CategorySearchComponent } from './category-search/category-search.component';
+import { CategoryDetailComponent } from './categories/category-detail/category-detail.component';
+import { CategorySearchComponent } from './categories/category-search/category-search.component';
 import { FoodComponent } from './food/food.component';
 import { FoodDetailComponent } from './food-detail/food-detail.component';
