+++ frontend/src/app/categories/categories.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { CategoriesRoutingModule } from './categories-routing.module';
+import { CategoryListComponent } from './category-list/category-list.component';
+import { CategoryDetailComponent } from './category-detail/category-detail.component';
+import { CategorySearchComponent } from './category-search/category-search.component';
+
+@NgModule({
+  declarations: [CategoryListComponent, CategoryDetailComponent, CategorySearchComponent],
+  imports: [
+    CommonModule,
+    CategoriesRoutingModule
+  ]
+})
+export class CategoriesModule { }
