+++ frontend/src/app/categories/category-detail/category-detail.component.html
