+++ frontend/src/app/categories/category-detail/category-detail.component.spec.ts
