+++ frontend/src/app/categories/category-list/category-list.component.css
