+++ frontend/src/app/categories/category-list/category-list.component.spec.ts
