+++ frontend/src/app/categories/category-search/category-search.component.html
