+++ frontend/src/app/categories/category-search/category-search.component.spec.ts
